Add User-Agent to fetch requests (#26)
* Adding User-Agent header to fetch requests

I noticed that certain URLs were throwing 403s when trying to fetch URLs.
The reason is that the page was intercepting the User-Agent for urllib
"Python-urllib/2.7" and then blocking it. I think it makes sense to identify
our preview bot with it's own User Agent. This shouldn't break any existing
functionality, but should allow us to preview URLS for cricket.

* Update model.py